+++ src/IlluminaSampleSheet/V1/NovaSeqSample.php
@@ -1,8 +1,8 @@
 <?php declare(strict_types=1);
 
-namespace MLL\Utils\IlluminaSampleSheet;
+namespace MLL\Utils\IlluminaSampleSheet\V1;
 
-class MiSeqSample extends Sample
+class NovaSeqSample extends Sample
 {
     public string $sampleId;
 
@@ -12,8 +12,6 @@ class MiSeqSample extends Sample
 
     public string $sampleWell;
 
-    public string $sampleProject;
-
     public string $i7IndexId;
 
     public string $index;
@@ -22,26 +20,32 @@ class MiSeqSample extends Sample
 
     public string $index2;
 
+    public string $sampleProject;
+
+    public string $description;
+
     public function __construct(
         string $sampleId,
         string $sampleName,
         string $samplePlate,
         string $sampleWell,
-        string $sampleProject,
         string $i7IndexId,
         string $index,
         string $i5IndexId,
-        string $index2
+        string $index2,
+        string $sampleProject,
+        string $description
     ) {
         $this->sampleId = $this->validateSampleId($sampleId);
         $this->sampleName = $this->validateSampleName($sampleName);
         $this->samplePlate = $samplePlate;
         $this->sampleWell = $sampleWell;
-        $this->sampleProject = $sampleProject;
         $this->i7IndexId = $i7IndexId;
         $this->index = $this->validateIndex($index);
         $this->i5IndexId = $i5IndexId;
         $this->index2 = $this->validateIndex($index2);
+        $this->sampleProject = $sampleProject;
+        $this->description = $description;
     }
 
     public function toString(): string
@@ -51,11 +55,12 @@ class MiSeqSample extends Sample
             $this->sampleName,
             $this->samplePlate,
             $this->sampleWell,
-            $this->sampleProject,
             $this->i7IndexId,
             $this->index,
             $this->i5IndexId,
             $this->index2,
+            $this->sampleProject,
+            $this->description,
         ]);
     }
 }
