+++ src/IlluminaSampleSheet/V1/NovaSeqXpSample.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace MLL\Utils\IlluminaSampleSheet;
+namespace MLL\Utils\IlluminaSampleSheet\V1;
 
 class NovaSeqXpSample extends Sample
 {
