+++ tests/IlluminaSampleSheet/NovaSeqSampleSheetTest.php
@@ -2,23 +2,19 @@
 
 namespace MLL\Utils\Tests\IlluminaSampleSheet;
 
-use MLL\Utils\IlluminaSampleSheet\DataSection;
-use MLL\Utils\IlluminaSampleSheet\NovaSeqDataSection;
-use MLL\Utils\IlluminaSampleSheet\NovaSeqHeaderSection;
-use MLL\Utils\IlluminaSampleSheet\NovaSeqSample;
-use MLL\Utils\IlluminaSampleSheet\NovaSeqSampleSheet;
-use MLL\Utils\IlluminaSampleSheet\ReadsSection;
-use MLL\Utils\IlluminaSampleSheet\SettingsSection;
+use MLL\Utils\IlluminaSampleSheet\V1\NovaSeqSample;
+use MLL\Utils\IlluminaSampleSheet\V1\NovaSeqSampleSheet;
+use MLL\Utils\IlluminaSampleSheet\V1\SettingsSection;
 use PHPUnit\Framework\TestCase;
 
 class NovaSeqSampleSheetTest extends TestCase
 {
     public function testNovaSeqSampleSheetAddsSectionsOnConstruction(): void
     {
-        $headerSection = $this->createMock(NovaSeqHeaderSection::class);
-        $readsSection = $this->createMock(ReadsSection::class);
-        $settingsSection = $this->createMock(SettingsSection::class);
-        $dataSection = $this->createMock(DataSection::class);
+        $headerSection = $this->createMock(\MLL\Utils\IlluminaSampleSheet\V1\NovaSeqHeaderSection::class);
+        $readsSection = $this->createMock(\MLL\Utils\IlluminaSampleSheet\V1\ReadsSection::class);
+        $settingsSection = $this->createMock(\MLL\Utils\IlluminaSampleSheet\V1\SettingsSection::class);
+        $dataSection = $this->createMock(\MLL\Utils\IlluminaSampleSheet\V1\DataSection::class);
 
         $novaSeqSampleSheet = new NovaSeqSampleSheet($headerSection, $readsSection, $settingsSection, $dataSection);
 
@@ -30,7 +26,7 @@ class NovaSeqSampleSheetTest extends TestCase
 
     public function testNovaSeqStandardSampleSheetToStringReturnsExpectedResult(): void
     {
-        $headerSection = new NovaSeqHeaderSection(
+        $headerSection = new \MLL\Utils\IlluminaSampleSheet\V1\NovaSeqHeaderSection(
             '4',
             'DonalDuck',
             'MyExperiment',
@@ -42,16 +38,16 @@ class NovaSeqSampleSheetTest extends TestCase
             'MyChemistry',
         );
 
-        $readsSection = new ReadsSection(101, 101);
+        $readsSection = new \MLL\Utils\IlluminaSampleSheet\V1\ReadsSection(101, 101);
 
-        $dataSection = new NovaSeqDataSection();
+        $dataSection = new \MLL\Utils\IlluminaSampleSheet\V1\NovaSeqDataSection();
 
         $dataSection->addSample(new NovaSeqSample('1', 'Sample-001-M001', 'RunXXXX-PLATE', '', 'UDP0090', 'TCAGGCTT', 'UDP0090', 'ATCATGCG', 'RunXXXX-PROJECT', 'description'));
         $dataSection->addSample(new NovaSeqSample('2', 'Sample-002-M002', 'RunXXXX-PLATE', '', 'UDP0091', 'CCTTGTAG', 'UDP0091', 'CCTTGGAA', 'RunXXXX-PROJECT', 'description'));
         $dataSection->addSample(new NovaSeqSample('3', 'Sample-003-M003', 'RunXXXX-PLATE', '', 'UDP0092', 'GAACATCG', 'UDP0092', 'TCGACAAG', 'RunXXXX-PROJECT', 'description'));
 
         $settings = new SettingsSection('AGATCGGAAGAGCACACGTCTGAACTCCAGTCA', 'AGATCGGAAGAGCGTCGTGTAGGGAAAGAGTGT');
-        $miSeqSampleSheet = new NovaSeqSampleSheet($headerSection, $readsSection, $settings, $dataSection);
+        $miSeqSampleSheet = new \MLL\Utils\IlluminaSampleSheet\V1\NovaSeqSampleSheet($headerSection, $readsSection, $settings, $dataSection);
 
         $expected = '[Header]
 IEMFileVersion,4
