+++ tests/IlluminaSampleSheet/ReadsSectionTest.php
@@ -2,14 +2,13 @@
 
 namespace MLL\Utils\Tests\IlluminaSampleSheet;
 
-use MLL\Utils\IlluminaSampleSheet\ReadsSection;
 use PHPUnit\Framework\TestCase;
 
 class ReadsSectionTest extends TestCase
 {
     public function testReadsSectionToStringReturnsCorrectFormat(): void
     {
-        $readsSection = new ReadsSection(150, 50);
+        $readsSection = new \MLL\Utils\IlluminaSampleSheet\V1\ReadsSection(150, 50);
 
         self::assertSame("[Reads]\n150\n50", $readsSection->toString());
     }
