+++ tests/IlluminaSampleSheet/SettingsSectionTest.php
@@ -2,14 +2,13 @@
 
 namespace MLL\Utils\Tests\IlluminaSampleSheet;
 
-use MLL\Utils\IlluminaSampleSheet\SettingsSection;
 use PHPUnit\Framework\TestCase;
 
 class SettingsSectionTest extends TestCase
 {
     public function testSettingsSectionToStringReturnsCorrectFormat(): void
     {
-        $settingsSection = new SettingsSection('AGATCG', 'TCG');
+        $settingsSection = new \MLL\Utils\IlluminaSampleSheet\V1\SettingsSection('AGATCG', 'TCG');
 
         self::assertSame("[Settings]\nAdapter,AGATCG\nAdapterRead2,TCG", $settingsSection->toString());
     }
